+++ src/shapes/Shape.h
@@ -7,7 +7,7 @@
 class Shape
 {
 public:
-    virtual bool hit(const Ray& r, double tMin, double tMax, HitRecord& hitRecord) = 0;
+    virtual bool hit(const Ray& r, float tMin, float tMax, HitRecord& hitRecord) = 0;
     virtual bool UpdateMaterial(Material* pMaterial) { _pMaterial = pMaterial; return true; }
     virtual std::string toString() = 0;
